+++ src/gt4py/backend/__init__.py
@@ -25,8 +25,6 @@ from .base import (
     from_name,
     register,
 )
-from .debug_backend import DebugBackend
-from .gt_backends import GTCUDABackend, GTMCBackend, GTX86Backend
 from .gtc_backend import (
     GTCCudaBackend,
     GTCDaceBackend,
@@ -36,4 +34,3 @@ from .gtc_backend import (
     GTCNumpyBackend,
 )
 from .module_generator import BaseModuleGenerator
-from .numpy_backend import NumPyBackend
