+++ tests/test_unittest/test_gtc/gtir_utils.py
@@ -14,10 +14,11 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from typing import List
+from typing import Any, Dict, List, Optional
 
 import factory
 
+from eve import Str
 from gtc import common, gtir
 
 from .common_utils import CartesianOffsetFactory, identifier, undefined_symbol_list
@@ -148,3 +149,9 @@ class StencilFactory(factory.Factory):
     name = identifier(gtir.Stencil)
     vertical_loops = factory.List([factory.SubFactory(VerticalLoopFactory)])
     params = undefined_symbol_list(lambda name: FieldDeclFactory(name=name), "vertical_loops")
+    api_signature = factory.LazyAttribute(
+        lambda x: [gtir.Argument(name=p.name, is_keyword=False, default="") for p in x.params]
+    )
+    externals: Dict[str, Any] = {}
+    sources: Optional[Dict[str, str]] = None
+    docstring: Str = ""
