+++ tests/test_unittest/test_gtc/test_dace.py
@@ -18,7 +18,6 @@ from copy import deepcopy
 
 import pytest
 
-from gt4py.backend.gtc_backend.defir_to_gtir import DefIRToGTIR
 from gt4py.definitions import BuildOptions
 from gt4py.frontend.gtscript_frontend import GTScriptFrontend
 from gtc.common import AxisBound, DataType
@@ -45,11 +44,10 @@ def stencil_def_to_oir(stencil_def, externals):
     build_options = BuildOptions(
         name=stencil_def.__name__, module=__name__, rebuild=True, backend_opts={}, build_info=None
     )
-    definition_ir = GTScriptFrontend.generate(
-        stencil_def, externals=externals, options=build_options
-    )
-    gtir = GtirPipeline(DefIRToGTIR.apply(definition_ir)).full()
-    return GTIRToOIR().visit(gtir)
+    gtir_stencil = GtirPipeline(
+        GTScriptFrontend.generate(stencil_def, externals=externals, options=build_options)
+    ).full()
+    return GTIRToOIR().visit(gtir_stencil)
 
 
 @pytest.mark.parametrize("stencil_name", stencil_registry.keys())
