+++ tests/test_unittest/test_gtc/test_gtir.py
@@ -17,20 +17,8 @@
 import pytest
 from pydantic.error_wrappers import ValidationError
 
-from eve import SourceLocation
-from gtc.common import ArithmeticOperator, ComparisonOperator, DataType, LevelMarker, LoopOrder
-from gtc.gtir import (
-    AxisBound,
-    Decl,
-    Expr,
-    FieldAccess,
-    FieldDecl,
-    Interval,
-    ParAssignStmt,
-    Stencil,
-    Stmt,
-    VerticalLoop,
-)
+from gtc.common import ArithmeticOperator, ComparisonOperator, DataType
+from gtc.gtir import Decl, Expr, Stmt
 
 from .gtir_utils import (
     BinaryOpFactory,
@@ -51,69 +39,18 @@ ANOTHER_ARITHMETIC_TYPE = DataType.INT32
 A_ARITHMETIC_OPERATOR = ArithmeticOperator.ADD
 
 
-@pytest.fixture
-def copy_assign():
-    yield ParAssignStmt(
-        loc=SourceLocation(line=3, column=2, source="copy_gtir"),
-        left=FieldAccess.centered(
-            name="foo", loc=SourceLocation(line=3, column=1, source="copy_gtir")
-        ),
-        right=FieldAccess.centered(
-            name="bar", loc=SourceLocation(line=3, column=3, source="copy_gtir")
-        ),
-    )
-
-
-@pytest.fixture
-def interval(copy_assign):
-    yield Interval(
-        loc=SourceLocation(line=2, column=11, source="copy_gtir"),
-        start=AxisBound(level=LevelMarker.START, offset=0),
-        end=AxisBound(level=LevelMarker.END, offset=0),
-    )
-
-
-@pytest.fixture
-def copy_v_loop(copy_assign, interval):
-    yield VerticalLoop(
-        loc=SourceLocation(line=2, column=1, source="copy_gtir"),
-        loop_order=LoopOrder.FORWARD,
-        interval=interval,
-        body=[copy_assign],
-        temporaries=[],
-    )
-
-
-@pytest.fixture
-def copy_computation(copy_v_loop):
-    yield Stencil(
+def test_copy():
+    copy_computation = StencilFactory(
         name="copy_gtir",
-        loc=SourceLocation(line=1, column=1, source="copy_gtir"),
-        params=[
-            FieldDecl(
-                name="foo",
-                dtype=DataType.FLOAT32,
-                dimensions=(True, True, True),
-            ),
-            FieldDecl(
-                name="bar",
-                dtype=DataType.FLOAT32,
-                dimensions=(True, True, True),
-            ),
-        ],
-        vertical_loops=[copy_v_loop],
+        vertical_loops__0=VerticalLoopFactory(
+            body__0__left__name="foo", body__0__right__name="bar"
+        ),
     )
-
-
-def test_copy(copy_computation):
     assert copy_computation
-    assert copy_computation.param_names == ["foo", "bar"]
+    assert set(copy_computation.param_names) == {"foo", "bar"}
 
 
-@pytest.mark.parametrize(
-    "invalid_node",
-    [Decl, Expr, Stmt],
-)
+@pytest.mark.parametrize("invalid_node", [Decl, Expr, Stmt])
 def test_abstract_classes_not_instantiatable(invalid_node):
     with pytest.raises(TypeError):
         invalid_node()
