+++ tests/test_unittest/test_layouts.py
@@ -14,7 +14,11 @@
 #
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-from gt4py.backend.gt_backends import make_cuda_layout_map, make_mc_layout_map, make_x86_layout_map
+from gt4py.backend.gtc_backend.base import (
+    make_cuda_layout_map,
+    make_mc_layout_map,
+    make_x86_layout_map,
+)
 
 
 def test_x86_layout():
